+++ test/Digipolis.Auth.UnitTests/Jwt/JwtBearerOptionsFactoryTests.cs
@@ -1,6 +1,8 @@
 ﻿using Digipolis.Auth.Jwt;
 using Digipolis.Auth.Options;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Http;
 using Microsoft.IdentityModel.Tokens;
 using Moq;
 using System;
@@ -29,7 +31,8 @@ namespace Digipolis.Auth.UnitTests.Jwt
 
             var jwtBearerOptionsFactory = new JwtBearerOptionsFactory(tokenValidationParametersFactory.Object, _testLogger);
 
-            var options = jwtBearerOptionsFactory.Create();
+            var options = new JwtBearerOptions();
+            jwtBearerOptionsFactory.Setup(options);
 
             tokenValidationParametersFactory.Verify(m => m.Create(), Times.Once);
             Assert.Same(tokenValidationParameters, options.TokenValidationParameters);
@@ -41,9 +44,13 @@ namespace Digipolis.Auth.UnitTests.Jwt
             var tokenValidationParametersFactory = new Mock<ITokenValidationParametersFactory>();
             var jwtBearerOptionsFactory = new JwtBearerOptionsFactory(tokenValidationParametersFactory.Object, _testLogger);
 
-            var options = jwtBearerOptionsFactory.Create();
+            var options = new JwtBearerOptions();
+            jwtBearerOptionsFactory.Setup(options);
 
-            var context = new AuthenticationFailedContext(null, options);
+            var mockHandler = Mock.Of<IAuthenticationHandler>();
+            var mockHttpContext = Mock.Of<HttpContext>();
+
+            var context = new AuthenticationFailedContext(mockHttpContext, new AuthenticationScheme("", "", mockHandler.GetType()), options);
             context.Exception = new Exception("exceptiondetail");
 
             await options.Events.AuthenticationFailed(context);
@@ -52,21 +59,5 @@ namespace Digipolis.Auth.UnitTests.Jwt
             Assert.Contains("exceptiondetail", _testLogger.LoggedMessages[0]);
         }
 
-        [Fact]
-        public async Task EmptyAuthenticationTicketIsSetWHenAuthenticationFailed()
-        {
-            var tokenValidationParametersFactory = new Mock<ITokenValidationParametersFactory>();
-            var jwtBearerOptionsFactory = new JwtBearerOptionsFactory(tokenValidationParametersFactory.Object, _testLogger);
-
-            var options = jwtBearerOptionsFactory.Create();
-
-            var context = new AuthenticationFailedContext(null, options);
-            context.Exception = new Exception("exceptiondetail");
-
-            await options.Events.AuthenticationFailed(context);
-
-            Assert.True(context.HandledResponse);
-            Assert.NotNull(context.Ticket);
-        }
     }
 }