+++ test/Digipolis.Auth.UnitTests/Services/AuthServiceTests.cs
@@ -1,8 +1,8 @@
 ﻿using Digipolis.Auth.Jwt;
 using Digipolis.Auth.Options;
 using Digipolis.Auth.Services;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Authentication;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Routing;
 using Microsoft.Extensions.Options;
@@ -20,19 +20,29 @@ namespace Digipolis.Auth.UnitTests.Services
         [Fact]
         public void ThrowsExceptionIfHttpContextAccessorIsNull()
         {
-            Assert.Throws<ArgumentNullException>(() => new AuthService(null, Mock.Of<ITokenRefreshAgent>(), Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>()));
+            Assert.Throws<ArgumentNullException>(() => new AuthService(null, Mock.Of<ITokenRefreshAgent>(), 
+                Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>(), Mock.Of<IAuthenticationService>()));
         }
 
         [Fact]
         public void ThrowsExceptionIfTokenRefreshAgentIsNull()
         {
-            Assert.Throws<ArgumentNullException>(() => new AuthService(Mock.Of<IHttpContextAccessor>(), null, Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>()));
+            Assert.Throws<ArgumentNullException>(() => new AuthService(Mock.Of<IHttpContextAccessor>(), null, 
+                Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>(), Mock.Of<IAuthenticationService>()));
         }
 
         [Fact]
         public void ThrowsExceptionIfUrlHelperFactoryIsNull()
         {
-            Assert.Throws<ArgumentNullException>(() => new AuthService(Mock.Of<IHttpContextAccessor>(), Mock.Of<ITokenRefreshAgent>(), null, Mock.Of<IOptions<AuthOptions>>()));
+            Assert.Throws<ArgumentNullException>(() => new AuthService(Mock.Of<IHttpContextAccessor>(), Mock.Of<ITokenRefreshAgent>(), 
+                null, Mock.Of<IOptions<AuthOptions>>(), Mock.Of<IAuthenticationService>()));
+        }
+
+        [Fact]
+        public void ThrowsExceptionIfAuthenticationServiceIsNull()
+        {
+            Assert.Throws<ArgumentNullException>(() => new AuthService(Mock.Of<IHttpContextAccessor>(), Mock.Of<ITokenRefreshAgent>(),
+                Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>(), null));
         }
 
         [Fact]
@@ -42,8 +52,8 @@ namespace Digipolis.Auth.UnitTests.Services
             var mockHttpContextAccessor = new Mock<IHttpContextAccessor>();
             mockHttpContextAccessor.SetupGet(m => m.HttpContext.User)
                 .Returns(user);
-            var authService = new AuthService(mockHttpContextAccessor.Object, Mock.Of<ITokenRefreshAgent>(), Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>());
-
+            var authService = new AuthService(mockHttpContextAccessor.Object, Mock.Of<ITokenRefreshAgent>(), Mock.Of<IUrlHelperFactory>(), Mock.Of<IOptions<AuthOptions>>(), Mock.Of<IAuthenticationService>());
+            
             var returnedUser = authService.User;
 
             Assert.Same(user, returnedUser);
@@ -66,7 +76,7 @@ namespace Digipolis.Auth.UnitTests.Services
             mockHttpContextAccessor.SetupGet(c => c.HttpContext).Returns(mockHttpContext.Object);
             var authOptions = new AuthOptions();
             var mockTokenRefreshAgent = new Mock<ITokenRefreshAgent>();
-            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenRefreshAgent.Object, Mock.Of<IUrlHelperFactory>(), Options.Create(authOptions));
+            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenRefreshAgent.Object, Mock.Of<IUrlHelperFactory>(), Options.Create(authOptions), Mock.Of<IAuthenticationService>());
 
             var returnedUserToken = authService.UserToken;
 
@@ -90,7 +100,7 @@ namespace Digipolis.Auth.UnitTests.Services
             mockHttpContextAccessor.SetupGet(c => c.HttpContext).Returns(mockHttpContext.Object);
             var authOptions = new AuthOptions { JwtTokenSource = "session" };
             var mockTokenRefreshAgent = new Mock<ITokenRefreshAgent>();
-            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenRefreshAgent.Object, Mock.Of<IUrlHelperFactory>(), Options.Create(authOptions));
+            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenRefreshAgent.Object, Mock.Of<IUrlHelperFactory>(), Options.Create(authOptions), Mock.Of<IAuthenticationService>());
 
             var returnedUserToken = authService.UserToken;
 
@@ -117,7 +127,7 @@ namespace Digipolis.Auth.UnitTests.Services
             mockHttpContextAccessor.SetupGet(c => c.HttpContext).Returns(mockHttpContext.Object);
             var authOptions = new AuthOptions { JwtTokenSource = "header" };
             var mockTokenRefreshAgent = new Mock<ITokenRefreshAgent>();
-            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenRefreshAgent.Object, Mock.Of<IUrlHelperFactory>(), Options.Create(authOptions));
+            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenRefreshAgent.Object, Mock.Of<IUrlHelperFactory>(), Options.Create(authOptions), Mock.Of<IAuthenticationService>());
 
             var returnedUserToken = authService.UserToken;
 
@@ -127,24 +137,19 @@ namespace Digipolis.Auth.UnitTests.Services
         [Fact]
         public async Task LogOutAsync()
         {
-            var mockAuthenticationManager = new Mock<AuthenticationManager>();
+            var mockAuthenticationService = new Mock<IAuthenticationService>();
             var mockHttpContext = new Mock<HttpContext>();
-            mockHttpContext.SetupGet(m => m.Authentication)
-                .Returns(mockAuthenticationManager.Object);
 
             var mockRequest = new Mock<HttpRequest>();
-            mockRequest.SetupGet(m => m.Scheme)
-                .Returns("http");
+            mockRequest.SetupGet(m => m.Scheme).Returns("http");
 
-            mockHttpContext.SetupGet(m => m.Request)
-                .Returns(mockRequest.Object);
+            mockHttpContext.SetupGet(m => m.Request).Returns(mockRequest.Object);
 
             mockHttpContext.SetupGet(m => m.User)
                 .Returns(new ClaimsPrincipal(new ClaimsIdentity(new Claim[] { new Claim("http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name", "user123") })));
 
             var mockHttpContextAccessor = new Mock<IHttpContextAccessor>();
-            mockHttpContextAccessor.SetupGet(m => m.HttpContext)
-                .Returns(mockHttpContext.Object);
+            mockHttpContextAccessor.SetupGet(m => m.HttpContext).Returns(mockHttpContext.Object);
 
             var mockTokenAgent = new Mock<ITokenRefreshAgent>();
             mockTokenAgent.Setup(a => a.LogoutTokenAsync("user123", "logoutUrl"))
@@ -165,13 +170,13 @@ namespace Digipolis.Auth.UnitTests.Services
             mockUrlHelperFactory.Setup(m => m.GetUrlHelper(It.IsAny<ActionContext>()))
                 .Returns(mockUrlHelper.Object);
 
-            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenAgent.Object, mockUrlHelperFactory.Object, Mock.Of<IOptions<AuthOptions>>());
+            var authService = new AuthService(mockHttpContextAccessor.Object, mockTokenAgent.Object, mockUrlHelperFactory.Object, Mock.Of<IOptions<AuthOptions>>(), mockAuthenticationService.Object);
 
             var controllerContext = new Mock<ControllerContext>();
 
             var result = await authService.LogOutAsync(new Microsoft.AspNetCore.Mvc.ControllerContext(), "Home", "Index");
 
-            mockAuthenticationManager.Verify(m => m.SignOutAsync(AuthSchemes.CookieAuth), Times.Once);
+            mockAuthenticationService.Verify(m => m.SignOutAsync(mockHttpContext.Object, AuthSchemes.CookieAuth, It.IsAny<AuthenticationProperties>()), Times.Once);
             Assert.Equal("logoutUrl", result);
             Assert.Equal("Home", urlActionContext.Controller);
             Assert.Equal("Index", urlActionContext.Action);